READ DESCRIPTION
Hey kashif i got it to write the name of the world to a file called
World saves
Now if you run the game it will crash
the reason is because i have hooke dup our level method to the random
number to actually get us our world picture. The reason this doesnt work
is because it takes a while to generate a picture so the level has
nothing to load while it si called so u get a illegal input == nul
because no input has been created. We have to movre all the level
methods  behind our world generation stuff u do that im busy over this
weekend